rename archive-storybook scripts file
