+++ packages/playwright/src/bin/build-archive-storybook.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-import { buildArchiveStorybook } from '@chromaui/shared-e2e/archive-storybook/scripts';
+import { buildArchiveStorybook } from '@chromaui/shared-e2e/archive-storybook';
 import path from 'path';
 
 // Discard first two entries (exec path and file path)
