+++ packages/shared/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig((options) => ({
-  entry: ['src/index.ts', 'src/archive-storybook/scripts.ts', 'src/util/filePaths.ts'],
+  entry: ['src/index.ts', 'src/archive-storybook/index.ts', 'src/util/filePaths.ts'],
   splitting: false,
   minify: !options.watch,
   format: ['cjs', 'esm'],
